+++ src/cpac/helpers/cpac_read_crash.py
@@ -1,11 +1,13 @@
 #!/usr/bin/env python
+'''cpac_read_crash.py
+
+`cpac_read_crash` is intended to be run inside a C-PAC container.
+The current run environment does not include the package `nipype`.
+'''
 
 import os
 import re
 
-from nipype import __version__ as nipype_version
-from nipype.utils.filemanip import loadcrash
-from traits.trait_errors import TraitError
 from sys import argv
 
 path_regex = re.compile(
@@ -13,7 +15,22 @@ path_regex = re.compile(
 )
 
 
+class NoNipype(ModuleNotFoundError):
+    def __init__(self, msg=None, *args, **kwargs):
+        no_nipype_message = __doc__.lstrip('cpac_read_crash.py').lstrip()
+        self.msg = '\n'.join([
+            msg, no_nipype_message
+        ]) if msg is not None else no_nipype_message
+        self.args = (self.msg,)
+
+
 def read_crash(path, touch_dir=None):
+    try:
+        from nipype.utils.filemanip import loadcrash
+        from traits.trait_errors import TraitError
+    except ModuleNotFoundError:
+        raise NoNipype from None
+
     try:
         crash_report = '\n'.join(loadcrash(path).get('traceback', []))
         print(crash_report)
@@ -36,5 +53,9 @@ def _touch_trait_error_path(crash_message):
 
 
 if __name__ == '__main__':
+    try:
+        from nipype import __version__ as nipype_version
+    except ModuleNotFoundError:
+        raise NoNipype from None
     print(f'Nipype version {nipype_version}')
     read_crash(argv[1])