+++ tests/test_cpac.py
@@ -1,15 +1,19 @@
 import pytest
+import sys
 
 from contextlib import redirect_stdout
-from cpac.backends import Backends
 from io import StringIO, TextIOWrapper, BytesIO
+from unittest import mock
+
+from cpac.backends import Backends
+from cpac.__main__ import run
+from CONSTANTS import set_commandline_args
 
 
-@pytest.mark.parametrize('platform', ['docker', 'singularity'])
-def test_loading_message(platform):
+def test_loading_message(platform, tag):
     redirect_out = StringIO()
     with redirect_stdout(redirect_out):
-        loaded = Backends(platform)
+        loaded = Backends(platform, tag=tag)
     with_symbol = ' '.join([
         'Loading',
         loaded.platform.symbol,
@@ -20,9 +24,29 @@ def test_loading_message(platform):
     redirect_out = TextIOWrapper(
         BytesIO(), encoding='latin-1', errors='strict', write_through=True)
     with redirect_stdout(redirect_out):
-        loaded = Backends(platform)
+        loaded = Backends(platform, tag=tag)
     without_symbol = ' '.join([
         'Loading',
         loaded.platform.name
     ])
     assert without_symbol in redirect_out.buffer.getvalue().decode()
+
+
+@pytest.mark.parametrize('argsep', [' ', '='])
+def test_pull(argsep, capsys, platform=None, tag=None):
+    def run_test(argv):
+        with mock.patch.object(sys, 'argv', argv):
+            run()
+            captured = capsys.readouterr()
+            checkstring = f':{tag}' if tag is not None else ':latest'
+            assert(
+                checkstring in captured.out + captured.err
+            )
+
+    args = set_commandline_args(platform, tag, argsep)
+
+    # test args before command
+    run_test(f'cpac {args} pull'.split(' '))
+
+    # test args after command
+    run_test(f'cpac pull {args}'.split(' '))