+++ tests/test_cpac_crash.py
@@ -5,18 +5,22 @@ import sys
 from unittest import mock
 
 from cpac.__main__ import run
-from CONSTANTS import PLATFORM_ARGS
+from CONSTANTS import set_commandline_args
 
 
-@pytest.mark.parametrize('args', PLATFORM_ARGS)
-def test_cpac_crash(args, capsys):
+@pytest.mark.parametrize('argsep', [' ', '='])
+def test_cpac_crash(argsep, capsys, platform=None, tag=None):
+    args = set_commandline_args(platform, tag, argsep)
     crashfile = os.path.join(
         os.path.dirname(__file__), 'test_data', 'test_pickle.pklz'
     )
-    argv = (
-        f'cpac {args} crash {crashfile}'
-    ).split(' ')
+    argv = ['cpac', 'crash', crashfile]
+    argv = ' '.join([
+        w for w in ['cpac', args, 'crash', crashfile] if len(w)
+    ]).split(' ')
     with mock.patch.object(sys, 'argv', argv):
         run()
         captured = capsys.readouterr()
-        assert("MemoryError" in captured.out or "MemoryError" in captured.err)
+        assert(
+            "MemoryError" in captured.out or "MemoryError" in captured.err
+        )